+++ mType/environment/registry/builtin/SqrtFunction.cpp
@@ -18,7 +18,7 @@ namespace environment::registry::builtin
             {
                 return std::sqrt(value);
             }
-            else if constexpr (std::is_same_v<std::decay_t<decltype(value)>, int>)
+            else if constexpr (std::is_same_v<std::decay_t<decltype(value)>, int64_t>)
             {
                 return std::sqrt(static_cast<float>(value));
             }
