+++ mType/value/ValueTypeUtils.cpp
@@ -28,7 +28,7 @@ namespace value {
         }
 
         return std::visit([](const auto& v) -> ValueType {
-            if constexpr (std::is_same_v<std::decay_t<decltype(v)>, int>) {
+            if constexpr (std::is_same_v<std::decay_t<decltype(v)>, int64_t>) {
                 return ValueType::INT;
             } else if constexpr (std::is_same_v<std::decay_t<decltype(v)>, float>) {
                 return ValueType::FLOAT;
